Client/requirements.txt
NumPy==1.20.0
grpcio-tools==1.35.0
sounddevice==0.4.1
Kivy==2.0.0